refactor: move breadcrumbs context from contexts/ to components/context/
- Move breadcrumbs-context.tsx from contexts/ to components/context/
- Update import paths in app/layout.tsx, components/nav-breadcrumbs.tsx, and hooks/use-breadcrumbs.ts
- Reformat imports in nav-breadcrumbs.tsx and use-breadcrumbs.ts to multi-line format
