+++ app/layout.tsx
@@ -13,7 +13,7 @@ import { UploadThingProvider } from "@/components/providers/uploadthing-provider
 import { UpstashRealtimeProvider } from "@/components/providers/upstash-realtime-provider";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { Toaster } from "@/components/ui/sonner";
-import { BreadcrumbsProvider } from "@/contexts/breadcrumbs-context";
+import { BreadcrumbsProvider } from "@/components/context/breadcrumbs-context";
 import { GENERAL_BRANDING_IMG_URL } from "@/lib/constants";
 import "./globals.css";
